Fix Utf8View.

+++ AK/Utf8View.h
@@ -26,7 +26,6 @@
 
 #pragma once
 
-#include <AK/Span.h>
 #include <AK/StringView.h>
 #include <AK/Types.h>
 
@@ -70,7 +69,7 @@ public:
 
     ReadonlyBytes bytes() const { return { begin_ptr(), byte_length() }; }
 
-    size_t byte_length() const { return m_string.length(); }
+    int byte_length() const { return m_string.length(); }
     int byte_offset_of(const Utf8CodepointIterator&) const;
     Utf8View substring_view(int byte_offset, int byte_length) const;
     bool is_empty() const { return m_string.is_empty(); }
